create folder for JAANet, move JAANet, fix imports, remove old haarscascade and .pt files
